pyproject.toml
[project]
name = "ws"
version = "0.1.0"
description = "Causal-Similarity Medical Decision Support System (CS-MDSS) - A medical AI demo combining similarity retrieval, causal inference, and decision chain visualization"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "albumentations>=2.0.8",
    "medmnist>=3.0.2",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pydicom>=3.0.1",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "shap>=0.49.1",
    "timm>=1.0.22",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "xgboost>=3.1.2",
    "torchvision>=0.22.1",
    "pytorch-lightning>=2.6.0",
    "monai>=1.4.0",
    "faiss-cpu>=1.11.0",
    "streamlit>=1.45.0",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.2",
    "plotly>=6.1.2",
    "opencv-python>=4.11.0.86",
    "rich>=14.2.0",
    "tensorboardx>=2.6.4",
    "tensorboard>=2.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
